refactor: extract common user service from OAuth callbacks
- Create shared findOrCreateUser function in user-service.ts
- Remove duplicated database logic from Google and Apple OAuth callbacks
- Improve modularity and maintainability of user management code
- Maintain identical functionality while reducing code duplication
